refactoring: move `OrderQuery::orderByUuidOrUrlHashQuery` logic from project base to package
